+++ packages/babel-plugin-lingui-transform-react/test/fixtures/component-plural/actual.js
@@ -1,3 +1,5 @@
+import { Trans, Plural } from 'lingui-react';
+
 <Plural
   value={count}
   offset="1"