+++ packages/babel-plugin-lingui-transform-react/test/fixtures/component-trans-expression/expected.js
@@ -1,3 +1,5 @@
+import { Trans } from 'lingui-react';
+
 <Trans id="Hi, my name is {name}" values={{
   name: name
 }} />;