+++ packages/babel-plugin-lingui-transform-react/test/fixtures/component-trans-text/actual.js
@@ -1,3 +1,5 @@
+import { Trans } from 'lingui-react';
+
 <span>Without translation</span>;
 <Trans>Hello World</Trans>;
 <Trans id="Hello World">Hello World</Trans>;