+++ packages/babel-plugin-lingui-transform-react/test/fixtures/env-production-remove-defaults/expected.js
@@ -1,1 +1,3 @@
+import { Trans } from 'lingui-react';
+
 <Trans id="msg.hello" />;