+++ packages/babel-plugin-lingui-transform-react/test/fixtures/format-number/expected.js
@@ -1,3 +1,5 @@
+import { Trans, DateFormat, NumberFormat } from 'lingui-react';
+
 <Trans id="The answer is {value,number}" values={{
   value: value
 }} />;