+++ packages/babel-plugin-lingui-transform-react/test/fixtures/nested-components/actual.js
@@ -1,3 +1,5 @@
+import { Select, Plural } from 'lingui-react';
+
 <Select
   value={genderOfHost}
   female={