+++ packages/example-usecase/src/Usecases/ElementAttributes.js
@@ -1,12 +1,12 @@
 // @flow
 import React from 'react'
-import { WithI18n, Trans } from 'lingui-react'
+import { withI18n, Trans } from 'lingui-react'
 
 type ElementAttributesProps = {
   i18n: Object
 }
 
-class ElementAttributes extends React.Component {
+class ElementAttributes extends React.Component<ElementAttributesProps> {
   props: ElementAttributesProps
 
   render () {
@@ -28,4 +28,4 @@ class ElementAttributes extends React.Component {
   }
 }
 
-export default WithI18n()(ElementAttributes)
+export default withI18n()(ElementAttributes)