+++ packages/babel-plugin-lingui-transform-react/CHANGELOG.md
@@ -1,7 +1,43 @@
 # Change Log
 
 All notable changes to this project will be documented in this file.
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/lingui/js-lingui/compare/babel-plugin-lingui-transform-react@1.0.5-0...babel-plugin-lingui-transform-react@1.0.5) (2017-08-28)
+
+
+
+
+**Note:** Version bump only for package babel-plugin-lingui-transform-react
+
+<a name="1.0.4"></a>
+## [1.0.4](https://github.com/lingui/js-lingui/compare/babel-plugin-lingui-transform-react@1.0.3...babel-plugin-lingui-transform-react@1.0.4) (2017-08-22)
+
+
+
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/lingui/js-lingui/compare/babel-plugin-lingui-transform-react@1.0.2...babel-plugin-lingui-transform-react@1.0.3) (2017-08-22)
+
+
+### Bug Fixes
+
+* ignore files without lingui imports [#62](https://github.com/lingui/js-lingui/issues/62) ([2eaf901](https://github.com/lingui/js-lingui/commit/2eaf901))
+
+
+
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/lingui/js-lingui/compare/babel-plugin-lingui-transform-react@1.0.1...babel-plugin-lingui-transform-react@1.0.2) (2017-08-16)
+
+
+### Bug Fixes
+
+* Transform only components from lingui-react ([10dd20f](https://github.com/lingui/js-lingui/commit/10dd20f))
+
+
+
 
 <a name="1.0.1"></a>
 ## [1.0.1](https://github.com/lingui/js-lingui/compare/babel-plugin-lingui-transform-react@1.0.0...babel-plugin-lingui-transform-react@1.0.1) (2017-07-10)