+++ packages/example-usecase/CHANGELOG.md
@@ -1,9 +1,48 @@
 # Change Log
 
 All notable changes to this project will be documented in this file.
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-      <a name="1.0.0"></a>
+  <a name="1.0.5"></a>
+## [1.0.5](https://github.com/lingui/js-lingui/compare/example-usecase@1.0.4...example-usecase@1.0.5) (2017-08-28)
+
+
+
+
+**Note:** Version bump only for package example-usecase
+
+  <a name="1.0.4"></a>
+## [1.0.4](https://github.com/lingui/js-lingui/compare/example-usecase@1.0.4-0...example-usecase@1.0.4) (2017-08-28)
+
+
+
+
+**Note:** Version bump only for package example-usecase
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/lingui/js-lingui/compare/example-usecase@1.0.2...example-usecase@1.0.3) (2017-08-22)
+
+
+
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/lingui/js-lingui/compare/example-usecase@1.0.1...example-usecase@1.0.2) (2017-08-22)
+
+
+
+
+<a name="1.0.1"></a>
+## [1.0.1](https://github.com/lingui/js-lingui/compare/example-usecase@1.0.0...example-usecase@1.0.1) (2017-08-16)
+
+
+### Bug Fixes
+
+* Update message loading in usecase ([623fb18](https://github.com/lingui/js-lingui/commit/623fb18))
+
+
+
+
+<a name="1.0.0"></a>
 # 1.0.0 (2017-07-12)
 
 
@@ -45,7 +84,7 @@ messages/languageData
 
 
 
-      <a name=""></a>
+<a name=""></a>
 #  (2017-07-10)