+++ packages/lingui-cli/CHANGELOG.md
@@ -1,7 +1,56 @@
 # Change Log
 
 All notable changes to this project will be documented in this file.
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="1.2.1"></a>
+## [1.2.1](https://github.com/lingui/js-lingui/compare/lingui-cli@1.2.0...lingui-cli@1.2.1) (2017-08-28)
+
+
+### Bug Fixes
+
+* Make default ignore pattern cross-platform ([43b4fb1](https://github.com/lingui/js-lingui/commit/43b4fb1)), closes [#70](https://github.com/lingui/js-lingui/issues/70)
+
+
+
+
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/lingui/js-lingui/compare/lingui-cli@1.2.0-3...lingui-cli@1.2.0) (2017-08-28)
+
+
+### Features
+
+* Add another hint to extract command ([38e9132](https://github.com/lingui/js-lingui/commit/38e9132))
+
+
+
+
+<a name="1.1.0"></a>
+# [1.1.0](https://github.com/lingui/js-lingui/compare/lingui-cli@1.0.5...lingui-cli@1.1.0) (2017-08-22)
+
+
+### Bug Fixes
+
+* ignore files without lingui imports [#62](https://github.com/lingui/js-lingui/issues/62) ([2eaf901](https://github.com/lingui/js-lingui/commit/2eaf901))
+
+
+### Features
+
+* Allow country-specific locales ([587b7bd](https://github.com/lingui/js-lingui/commit/587b7bd))
+
+
+
+
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/lingui/js-lingui/compare/lingui-cli@1.0.4...lingui-cli@1.0.5) (2017-08-16)
+
+
+### Bug Fixes
+
+* Return empty string when fallbackLaguage is empty ([ea645ce](https://github.com/lingui/js-lingui/commit/ea645ce)), closes [#51](https://github.com/lingui/js-lingui/issues/51)
+
+
+
 
 <a name="1.0.4"></a>
 ## [1.0.4](https://github.com/lingui/js-lingui/compare/lingui-cli@1.0.3...lingui-cli@1.0.4) (2017-07-10)