+++ packages/lingui-conf/CHANGELOG.md
@@ -1,7 +1,26 @@
 # Change Log
 
 All notable changes to this project will be documented in this file.
-See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+<a name="0.9.1"></a>
+## [0.9.1](https://github.com/lingui/js-lingui/compare/lingui-conf@0.9.0...lingui-conf@0.9.1) (2017-08-28)
+
+
+### Bug Fixes
+
+* Make default ignore pattern cross-platform ([43b4fb1](https://github.com/lingui/js-lingui/commit/43b4fb1)), closes [#70](https://github.com/lingui/js-lingui/issues/70)
+
+
+
+
+<a name="0.9.0"></a>
+# [0.9.0](https://github.com/lingui/js-lingui/compare/lingui-conf@0.9.0-3...lingui-conf@0.9.0) (2017-08-28)
+
+
+
+
+**Note:** Version bump only for package lingui-conf
 
 <a name="0.8.1"></a>
 ## [0.8.1](https://github.com/lingui/js-lingui/compare/lingui-conf@0.8.0...lingui-conf@0.8.1) (2017-07-10)